Adding TRUTH5 derivation
This derivation is a test for measurements.  In principle, it contains
everything that they might need: the normal contents of TRUTH3 for
acceptance studies, Born leptons for Born-level calculations, and the
first two generations of the truth record for matrix-level reweighting
(depending on the generator, one generation may not be sufficient).
This is setup with H->ZZ in mind, but should be sufficient for any
measurement that we want to undertake!


Former-commit-id: 288055f5015191acfe2ce25e9d58ab34fcab89f9