cleanup and implement the suggestion by Joerg
Details:

reorder the parameters so that now they're accessed in the same order
as done for LAr in the current L1Topoconfig_Physics_pp_v7.xml.

Also change the behavior of the 'NumberLeading' parameter so that now
it behaves as @stelzer suggested, i.e. it only checks the first N
TOBs. We should discuss whether we want to apply the N cut before or
after the ET cut (I think that for the relevant cases 'only-the-first'
and 'all-tobs' it doesn't make a difference).


Former-commit-id: b96f950401ebdeab2f27e4b2e99e3f0a6121f4d2