Avoid using HOME-relative paths for configuration
The code looked for certain password and configuration files in paths relative
to the running user's home directory. In reality there is only one copy of
these files, and it's in ~atlidbs on AFS. There's no reason not to run the
scripts as a different user, and no reason why some of these files should be
different for each user. This change just hard-codes the paths to the files so
that we can finally stop prefixing commands with `HOME=~atlidbs` to get them to
work.


Former-commit-id: 6b1053abd95f5758dc900234a2ed3b51a5c90f00