StoreGate: VarHandleBase no longer derives from VarHandleKey.
The normal use of a handle is to construct it from a VarHandleKey.
We want this operation to be fast.
However, when VarHandleBase derives from VarHandleKey, we were copying
the contents of the key, which was observed to be a significant overhead.

Now, VarHandleBase no longer derives from VarHandleKey.
Instead, it contains a pointer to the key, which is initialized in the
constructor from the key.  We add forwarding methods for the VarHandleKey
methods that are needed in handles.

For the case where a handle is not initialized from a key, we create a new
VarHandleKey instance, referencing it via a unique_ptr.


Former-commit-id: a4f2ea17184d85c3db548b123e8ac6f47fe8c294