AthenaBaseComps: VarHandleBase no longer derives from VarHandleKey.
The normal use of a handle is to construct it from a VarHandleKey.
We want this operation to be fast.
However, when VarHandleBase derives from VarHandleKey, we were copying
the contents of the key, which was observed to be a significant overhead.

Now, VarHandleBase no longer derives from VarHandleKey.

Adjust property handling for this.


Former-commit-id: 4b43d68edfd1936421571f21bf3fcaad443f29d8