Adding bosons to Sherpa samples
In some cases, there is no real ambiguity about what type of particle
produced a set of leptons -- for example, in the case of an e+/e- pair,
in a V+jets sample, that pair must come from a Z/gamma*, and the only
ambiguity is about what we call it.  Similarly for the case of an e/nu_e
pair.  In some more complex cases, there is still no ambiguity, as in
the case of e+e-mu+mu- (must be two Zs, cannot be two Ws).

This commit finally adds functionality to the truth collection maker
that has been promised for quite some time.  It looks through the Sherpa
record to collect the status code 20 (new feature since Sherpa 2.2.2) or
status code 3 (if there are no status code 20) leptons.  In the case of
an odd number of those leptons, it complains.  In the case of an even
number of leptons, it checks on whether we are in one of those Q-M
unambiguous cases, and if we are then it adds the appropriate bosons to
a boson collection.  If it does not know what to do (e.g. if the case is
ambiguous, as in e+/e-/nu_e/nu_e) then it does not treat the event, so
no user may do something truly evil with the information.

At the moment, the code does not insert these bosons into the connected
truth tree, it only creates them and puts them into a truth collection.
It is unclear at the moment if it would be useful to have them in a
truth tree or not, but that can be left for a future commit.

This satisfies the last long-standing request of the truth derivations.


Former-commit-id: cf6765e52d3b79feda7fbabb3b343e53584674db