This branch fixes the incorrect parsing of the job's log file (ATLASJT-371)
This branch fixes the incorrect parsing of service names from a sub-step's
log file during log file validation. Long service names are sometimes
listed in the job summary section of the log file without a space
between the name and the character '|'. Before the fix the log parser
was identifying these lines from the summary as additional errors. With
the fix in this branch the parser ignores such lines.


Former-commit-id: a7d5cc4313ddaa11ff788ec1c75f689027124443