merging of strip and wire PRDs for sTGC (ATLASSIM-3459)
The PRDs build from sim hits during (sTGC) fast digitization
are inputs to quickly progress with reconstruction validation
for the NSW. With this merge request hits in strips and wires
get merged to one PRD which then can enter reco. In case of
PRDs for each of the readouts reco gets confused. Also the
error treatment for the merged PRDs is taken into account
properly. The actual PRDs based on full digitization output
may look different, but are still under design.


Former-commit-id: 8de2e71dcaaefc9886df06cb5376a4ab65807a6f