Check for potential NaN (ATLJETMET-700)
The new JMS calibrations optionally have a bin in terms of
log(m/Et) or log(m/E) or log (m/pT).  In any of these cases, the
logarithm can end up as NaN, most notably if m is zero, but also
potentially if m is so small and pT is so large that a float cannot
represent the value.  This shouldn't happen, but this commit adds
protections to watch for this.


Former-commit-id: ce12194eafd8697e6acb847049d8eccab2c172d3