CxxUtils: Adding some helper classes for concurrency.
The CachedPointer and CachedValue classes, used for managing a concurrent
cache of some value, are copied here from AthLinks, as they have no other
dependencies and may be useful for other things.

Add new class ConcurrentRangeMap, a map from sorted ranges to payload
objects, allowing for concurrent lockless reads.  This was written
with the conditions container use case in mind.



Former-commit-id: f9d15b4c88661c953c412b7fcc0b4a2ac840b475