StoreGate: Cache event store flag in the handle key.
VarHandleBase::storeFromHandle had a test to see if the referenced event
store was the primary event store.  This was based on the name in the
service handle, so this test was relatively expensive.  However, this is
a property of the key, so speed up this test by caching the result in the
HandleKey.

Former-commit-id: 15398905068eaaabc589777260baa42d393d331b