Modifications to ensure that tracks are properly contained within an Roi
Using simply etaMinus < eta < etaPlus is seriously flawed, since the Roi
can be extended in z, and as such tracks can be fully contained inside an
RoI even if it is apparently outside the eta range.

For instance, consider an RoI with |z| < 225 mm and with etaMinus = etaPlus = 0,
so it is parallel sided and has an eta "width" or 0, so no tracks could
be within this RoI.

Clearly this is nonsense - a tracks with z = -225 could have a whole range of
eta values from eta=0 to eta ~ 0.4 and still be fully contained.

These changes finally fix this properly for the offline code.


Former-commit-id: ff6e55e0061ff7c0a18a14bef38bee6e452c1f8c