SGComps: Fix input supercedence for xAOD types.
An earlier change added the feature in which having a WriteHandle for an
object would block that object from being read from disk.  However, the
logic didn't work quite correctly for the case when an xAOD object 
is being read from the input file and, in the same job, a non-xAOD object
of a different type but the same SG key is created.  The non-xAOD object
ended up blocking the read of the aux store for the xAOD object.

This led to an ATN failure in RecExRecoTest.
Now fixed.


Former-commit-id: ef8bf27a5aa21d248fcf9206d370cf166765a55d