Merge branch 'BMGChambers_20150501' into 'master'
Decoding and cabling for BMG chambers. As there were many
changes in master wrt the original code a manual cherry-pick
was needed. This one is the first of a series. The code
still has some bugs (but compiles fine) which will be fixed
by subsequent manual cherry-picks. It will be done
step-by-step to keep the history like the code went into
21.0/21.1 . It should not go into 21.3 because there the
release 21 style of decoding is used and in sync with 21.0.

See merge request !1726


Former-commit-id: e3714a2c8e571941e76ab13b55a3720c4dea7b3c