Partial migration of MuonCreatorAlg to use handles
Full migration of this algorithm is still progress (see ATLASRECTS-4108).
However, this partial migration is required to allow the code to run after the migration of the MuonCombinedMuonCandidateAlg.
Apologies for not checking the effects of that ahead of time: this problem was masked by a different issue that was preventing running, so I missed it.
This fix means that the retrieval of the MuonCandidateCollection is now done using a ReadHandle.
Additionally, the newest constructor is used for both ReadHandles currently present in the algorithm.
Since the Stau instance of the algorithm does not use the MuonCandidateCollection, the MuonCandidateCollection ReadHandle will not be created if the BuildSlowMuon flag is true.
This also required adding a couple of extra protections in the MuonCreatorTool.


Former-commit-id: aacfd1e71591d2ce974a02e99ac68a019c842385