Update of Cosmic Ray and Cavern Background Simulation Configuration
In ATLASSIM-3214 the naming of TrackRecord input/output file arguments was discussed
and it was agreed to use `inputEVNT_TRFile` and `outputEVNT_TRFile` for all
simulation jobs which either read in or write out TrackRecord files. These
new names were agreed with the AMI developers. (The old names having been too
long.)
This commit pulls in changes from `SimuJobTransforms-00-33-03`, `SimuJobTransforms-00-33-04`
and `SimuJobTransforms-00-33-05`.
It also pulls in corresponding updates to the syntax or RTT and ART jobs.

+	* python/simTrfArgs.py : Only input TR and output TR, not separate cosmic
+	/ cavern TR files
+	* python/SimTransformUtils.py : Adding a sub-step to deal with jobs that
+	write out TRs and jobs that read in TRs separately
+	* share/skeleton.EVGENtoHIT_ISF.py : updating to configure only based on
+	TR file requirements and whether beam config is cosmic, rather than using
+	type of TR file as a hint
+	* share/skeleton.EVGENtoHIT_MC12.py : updating to configure only based on
+	TR file requirements and whether beam config is cosmic, rather than using
+	type of TR file as a hint
+	* scripts/AtlasG4_tf.py : updating list of input arguments
+	* scripts/Sim_tf.py : updating list of input arguments
+	* Tagging: SimuJobTransforms-00-33-03


Former-commit-id: f015a48aa6ba345486dde39b1b1570459d236054