Move retrieval of PRDs for MuGirl to the algorithm
MuGirl builds its own segments, for which it requires PRDs.
At the moment, the retrieval of PRDs is handled by a tool, the MuonPrepRawDataCollectionProviderTool.
This retrieval is buried several tools deep in the algorithm, so there is no way for the scheduler to know about this dependence.
This commit changes this by having the MuonInDetExtensionAlg retrieve the PRD containers and pass them to the MuonInsideOutRecoTool.
Other functionality of the MuonPrepRawDataCollectionProviderTool, which retrieves the particular PRD collections that are needed for the segment reco, is migrated to the MuonInsideOutRecoTool.
The MuonLayerSegmentFinderTool will then take these collections from the MuonInsideOutRecoTool, rather than retrieving them itself using the MuonPrepRawDataCollectionProviderTool.
Its interface is modified to allow this.
As instances of the MuonCombinedInDetExtensionAlg are used by other algorithms which don't require PRDs, a flag is added to determine whether to retrieve the PRDs.
This flag is now added to the configuration of the MuonInsideOutRecoAlg.
A new interface is then added to the IMuonCombinedInDetExtensionTool that passes the PRD containers to the tool.
Other tools that use this interface need to have methods for this interface added.
By checking out the commit from MR 6839 I was able to confirm that q431 runs fine with these changes.
Mentioning ATLASRECTS-4237 for bookkeeping purposes.


Former-commit-id: 496e4f4ca29763e4d0be633b3a992c1107357500