change criteria from 'min<X<max' to 'min<=X<max'
Details:
This was suggested by @landon: the motivation is that in the past (for
example selecting some eta range from 0 to 1.6) we always had to tune
the threshold to some non-intuitive value (due to the quantization of
the TOB coordinates).
This choice should make things more intuitive.
Both menu and firmware people agree with this proposal
(indico.cern.ch/event/684166).

Former-commit-id: 50f4d8966e4153c42f910aabb7518e23fd886c07