Try to ensure that commands sent via the G4UImanager work as intended
In ATLASSIM-3454 it was noted that many commands added to
`simFlags.G4Commands` and then executed via the `G4UImanager` seem
to have no effect. This was determined to be due to the fact that
they were applied at the wrong point during the Geant4 initialization.
In order to address this, the code has been refactored to try the
commands at multiple points during the initialization until they
succeed. These changes have been checked to make sure that they
don't alter simulation output. They do result in more checks being
switched on the Geant4 than previously so additional `G4Exceptions`
appear in the log. Further iterations may be required.