Setting default hard scatter vertex
In some cases, generators don't record a 'hard scatter vertex' in the
HepMC record.  We have some code in TestHepMC to protect against such
cases, but it is possible to run truth derivations on EVNT files
produced before that protection was in place.  Moreover, I can imagine
that the protection as written might fail in very weird cases (these
really shouldn't happen, but one never knows).  Therefore, trying to
provide a default value for the hard scatter vertex in case it isn't
known (the first vertex).  This is the standard adopted in other parts
of the athena code base when this test goes south as well.


Former-commit-id: 4547a7467b8fad9e3f76b59ec16cd6c8e4e22d65