Move storegate retrieval away from the MuonClusterSegmentFinder
The MuonClusterSegmentFinder was retrieving PRDs itself, or using the MuonPrepRawDataCollectionProviderTool to do so.
These changes move all retrievals from Storegate to the algorithms, and switch to using dataHandles.
Interfaces to, and functions in, the MuonClusterSegmentFinder are updated to allow for passing collections from the algorithm, rather than retrieving them locally.
A new interface is added to account for the fact that the MooSegmentFinderAlg creates a vector of MuonPrepDataCollections, rather than using the MuonPrepDataContainer.
Finally, some functionality of the MuonPrepRawDataCollectionProviderTool, for retrieving only the PRDs in a certain detector region, is migrated to the MuonClusterSegmentFinder.
Mentioning ATLASRECTS-4237 for bookkeeping purposes.


Former-commit-id: 67ae7711c494e3e523261d66fcbacaab9259d690