Define lepton types kTriMuon and kTriElectron for ttZ likelihood
This now allows to set the lepton type to kTriMuon and kTriElectron to
deal with different final-state scenarios for ttZ. The ttZ->trilepton
likelihood will now only be called of one of these two cases is
set. The mixed-flavour cases can be reduced to standard ljets
reconstruction (ttbar only), because the SF lepton pair can be
associated with the Z boson without KLFitter reconstruction.


Former-commit-id: c3a558bf4b6a5a84e6e1c5c8b4b28f16590d348b