Update OutputWidthSelect* parameters
Details:
The new values reflect what's in
https://svnweb.cern.ch/trac/atl1calo/browser/firmware/L1Topo/trunk/processor/sources/common/libraries/L1TopoDataTypes.vhd#L51

It looks like these values were updated in the firmware in rev 2783
(Jun 2016), but that the change was not propagated to the software.
Thanks @nakahama for spotting this difference.

I tested this change with
`TrigConfTopoStandAlone L1Topoconfig_Physics_pp_v7.xml`
using several `inputdump.txt' files I had on disk,
and I don't see any difference in the simulated output bits.


Former-commit-id: 88da6637ce89ee139827b7c2b054e460b2463ab2