Add extra substructure moments to large-R trigger jets
Add do_substructure bool, currently hardcoded to True for large-R jets and False for all others,
to toggle the addition of a number of substructure moments. Not all of these moments wil be
taken forward to 21.1; we need to evaluate the CPU cost etc and take just cheapest and most useful.


Former-commit-id: 99e2cff7948e20d37d1d6a54d7157a7abfd76d49