Fixing the SuppressRedFactor in CaloMontoring webdisplay
The lack of definition of SuppressRedFactor in BinsDiffFromStripMedian algorithm
lead to inconsistencies between express and bulk due to the suppression associated
to the peak fo the 2D histogram.
But the histogram peak should not affect diagnosis in a different region.
See for instance: EMTopo barrel plots in processing 1 and 2 of run 339957/CosmicCalo


Former-commit-id: 7138ff1321262ba49ffecd9aa504f2f209653161