Fix for ATLASRECTS-3861
The DerivationFramework SlimmingHelper, which is used by a number of performance DPDs made at Tier-0, imports a python fragment that builds xAOD truth when it encounters the GEN_EVENT container. This fragment is normally used for TRUTH format making and is called by the slimming helper for adding certain truth branches. However, in performance DPDs the input is transient ESD, and the xAOD truth has already been run. So there is a conflict as the truth metadata is written and a crash, reported in ATLASRECTS-3861. This MR by-passes this issue by removing the import of the truth-building fragment. Note that in newer versions of the SlimmingHelper used in 21.2, this import was dropped anyway, but it is part of a much bigger set of changes that we aren't ready to import to 21.0 at this stage.


Former-commit-id: 7885d264ea0044f4a8dcd652464e75f74a994607