SGComps: Improve handling of ProxyProviderSvc initialization loops.
It's easy to get initialization loops involving ProxyProviderSvc, for example
if a provider tries to access StoreGate.  In such a case, preLoadProxies
may be called before ProxyProviderSvc is completely initialized.
This can result in some providers being skipped.

To improve this: in preLoadProxies, check to see if the PPS is initialized.
If not, then defer the operation until initialization has completed.

Fixes ctest failures observed in RegSelSvcTest and TrigUpgradeTest.



Former-commit-id: 5c5d733496d354c8bff5d93adb1315d84822d912