CaloDetDescr: Const accessors should return const pointers.
CaloDetDescrManager has a bunch of accessors that return pointers
to elements and descriptors.  These pointers should be const if the
accessors themselves are const.

In some cases, we add non-const accessors, returning non-const pointers.
The names of these accessors end in _nonconst, to make them easily
searchable.

CaloAlignTool is modifying the contents of a locked CaloDetDescrManager
object in SG in order to apply alignments.  Add a const_cast for now.
We should get rid of this once we switch to the new scheme for alignments.


Former-commit-id: e7bf83ae681ec73ec02c09e6476f97bf361781d5