Update previous merge request with a special treatment of whether to attach an forID RoI in BjetVtx instance. Given the forID is not visible after jet splitting in the second stage and thus the descriptor gets attached again with the name forID it proves easier to avoid this rather than deal with it in the analysis.
Modified TrigRoiUpdater.cxx to use the == operator
when comparing RoIs instead of comparing coordinates
of each RoI. Operator can also catch cases of
duplicate composite RoIs.
Also fixed a typo in
DetectorDescrition/RoiDescriptor/Root/RoiUtil.cxx:
In the definition of the == operator when comparing
roi0.zedMinus() and roi1.zedMinus()
the if statement required the difference to be
>1e-7 for the two rois to be identical. Changed
it to <1e-7.


Former-commit-id: fe38c23ef92aa17b48ef5113479e2faa12fc228b