AsgExampleTools: Disable ToolHandleTest::nonempty_tool in Athena.
ToolHandleTest::nonempty_tool never actually worked as intended on Athena,
since it relies on a feature of ToolHandle that does not exist in the Gaudi
version.  It was really only by chance that this test had been passing,
and a change in behavior of ToolHandle in Gaudi v30 is causing it to fail.
So comment out the body of this test in Athena.



Former-commit-id: 80a0c27ceb909210986e72d8028b3cf764520734