Remove redundant isSuccess calls in CHECK macros
Remove the redunant isSuccess call in CHECK(sc.isSuccess()) as the CHECK
macro already does this. In theory this means that RECOVERABLE errors
are now correctly propagated to the CHECK macro whereas before they were
masked. In practice this will make no difference as RECOVERABLE is not
being returned by any of the calls in question.


Former-commit-id: 7b037a4142dc34db283d26546de1eeb07899059f