StoreGate: Add CondHandleKey::start().
For performance, we really want to cache a pointer to the conditions container
in CondHandleKey, and CondHandleKey attempts to retrieve such a pointer
during initialize().  However, for conditions objects that are created
by CondInputLoader, this is too early, since CondInputLoader doesn't
create the objects until start().

So add a new virtual method to VarHandleKey that we can call during start().
Add an implementation for it in CondHandleKey that makes another attempt
to retrieve the conditions container.


Former-commit-id: 376abb87a020557dc5c0968191c1950b593347c3