AthenaKernel: Some const-correctness fixes.
The init functions registered by addInit were using a special, private
version of find() to retrieve a modifiable version of the BIB.
This was then problematic when addInit was used in AthContainers; this
had required a const_cast.

Change the interface to instead pass the (non-const) BIB pointer to the init
function.  In that case, was can also get rid of the private find_nc function.


Former-commit-id: ab19cfed9255d7cab6ebde92d7a9917d773aadc2