fix MS vertex reconstruction
In !7003 some developments for MS vertex reconstruction present
in master were manually sweeped back. They included some
migration to WriteHandles. It turned out that the record method
of WriteHandles is implemented differently in 21.3 than in master.
It does not automatically associate the aux store. This results
in runtime crashes. They are resolved by this MR.


Former-commit-id: a0c3f7f117b0f2d5f7e706a217de5587461a9690