Build MuonSystemExtensions in the MuonCombinedInDetCandidateAlg
Right now the MuGirlAlg and MuGirlStauAlg both separately build MuonSystemExtension objects from InDetCandidates.
(These are used to determine which layers of the Muon Spectrometer are intersected by extrapolation from an ID track.)
This is an unnecessary duplication of effort: with this change, the code now builds the MuonSystemExtensions once, at the same time as the InDetCandidates, which then cache the results.
The MuonInsideOutRecoTool and MuonStauRecoTool are updated to no longer use the MuonSystemExtensionTool.
The MuonSystemExtensionTool no longer saves the MuonSystemExtensions in StoreGate: instead, they are attached to InDetCandidates which are saved in SG.
Please see ATLASRECTS-4266 for more details.


Former-commit-id: 8727b49f48603b9abf5a80b9f0ece36fb4d2c92c