CaloDetDescr: Const iterators should yield const pointers.
CaloDetDescrManager has a number of const methods that return iterators
over elements and descriptors.  The pointers that one gets from these
iterators should also be const.

In some cases, we add non-const versions, that return iterators that return
non-const pointers.  The names of these end in _nonconst, to make them easily
searchable.

CaloSuperCellAlignTool is modifying the contents of a locked CaloDetDescrManager
object in SG in order to apply alignments.  Add a const_cast for now.
We should get rid of this once we switch to the new scheme for alignments.


Former-commit-id: 9e56c0d8326f8f665419e84590fa463c5aaac550