passing through L2 before going to EF for late muon triggers
Late muon triggers are up to now executing EF immediately on
top of L1 outputs. This causes high CPU costs due to ID tracking
in EF even if the event is rejected because of missing muon
information. Therefore an intermediate passing through L2muonSA
is implemented to reduce the number of events passed to EF.


Former-commit-id: b0203984fe58ae4ffcb7be8b84f06a8a4fac97c9