Fix for unretrieved ToolHandles in the muon domain
The unretrieved ToolHandles here fall into four categories.
Missing retrievals: these are added.
ToolHandles for tools that are not actually used: these are removed.
ToolHandles that are only retrieved if a flag is true: if the flag is false, the ToolHandle is disabled.
And, in one case, a ToolHandle that is retrieved in the finalize() method of the tool, probaby because the finalize() method appears before the initialize() method in the file.  The retrieval is moved to initialize().


Former-commit-id: 5adc3f9fcdc89f5a60b43410a7658004fc839d7b