StoreGate: Small improvements to VarHandleBase ctor/dtor.
Change VarHandleBase::setStoreFromHandle to return a bool rather than a StatusCode.
The StatusCode dtor was a hotspot in profiles when this was called
from the VarHandleBase constructor.

In ~VarHandleBase, only call resetProxy() if m_ownedKey is set.
This is strictly redundant, but it speeds up the common case
where m_ownedKey is not set.


Former-commit-id: 4e006fc0d3fe89acd104c7059de39cc5b6238311