Migration of MuonCreatorAlg to use datahandles (ATLASRECTS-4108, ATR-16944)
Migrated all collections in MuonCreatorAlg to use datahandles.
Also included all of the cluster making/recording into MuonCreatorAlg, and removed the calls to initialize/finalize clusters (which is what previously handled the clusters) from MuonCombinedRec_postprocessing.py and MuonCombinedRec_preprocessing.py
Added an option so that we can build SA muons with the creator alg (needed use case for the trigger).


Former-commit-id: 7a270200aaa4c5aac0eedfb8d88ebd3efd8b6383