Set AddSectors to be false by default
This option and the associated code were originally added as part of a sweep of NSW-related improvements from 21.3.
It leads to a considerable increase in the number of segments being reconstructed, but not in the number of tracks, suggesting the new segments are fakes.
Additionally, it is the top suspect for the increase in CPU usage reported in ATLASRECTS-4257.
If needed for NSW reconstruction tests, it can always be re-enabled in the configuration.


Former-commit-id: 09f7b5f5edcac445fd91c221c3519dc37cef0de3