Add new max radius cut specifically for layer EMB1 and EME1
During the Simulation workshop (Dec 2017) it was decided that we
should just merge hits within 5mm everywhere except in layers 1
(EMB1) and 5 (EME1) where only hits within 1mm should be merged.
This commit adds a new property to allow the max radius (squared)
cut to be configured separately for EMB1 and EME1.


Former-commit-id: 62b06329b5f1da88291fadbd7e572427cf504e98