Simplify merging scheme for LArFCS_StepInfoSD
Added a simplified version of the `update_map` method to
`LArFCS_StepInfoSD` implementing the recommendations from
the Dec 2017 Simulation Workshop. (1mm merging in EMB1 and
EME1 and 5mm elsewhere)
In order to minimise changes the old method is still in the
`FCS_StepInfoSD` base class.

Also updated the `ISF_FastCaloSimParametrization_SimPostInclude_1mm.py`
job options to configure the sensitive detectors according to
the recommendations.


Former-commit-id: 4194af942cbdf5c62c0adcf0ce3b689b920968b3