Remove use of MuonPatternSegmentAssociationTool
This tool provides an interface to the map between the muon segment combinations and muon pattern combinations.
This map is currently put into StoreGate: when needed, the tool retrieves it so it can be accessed or modified.
This is obviously not compatible with the use of DataHandles.
Additionally, it's not necessary, given that the map is only used in one algorithm, the MooSegmentFinderAlg.
Instead, the map is now a member of the top-level tool in the algorithm, the MooSegmentCombinationFinder, which passes it to sub-tools as needed.
The MuonPatternSegmentAssociationTool is thus completely removed from the algorithm and its sub-tools.
Interfaces to sub-tools are updated to include the map, which occasionally must also be further passed to methods called by the interface methods.
Methods of the MuonPatternSegmentAssociationTool are replaced by the equivalent methods of std::multimap.
The MuonHoughPatternFinder was only using the map for CSC segment-finding, but this is done in a different fashion now, so this method is removed completely.
(MuonHoughPatternFinder is anyway not used for reconstructing collision data any more.)


Former-commit-id: 352460a9f04074ef8ed7e8b96078a41a22f6bcdb