removal of AmdcData package
In the past this package was used to store the muon geometry
files under a version control system. With the imposed ATLAS
policy that the git repository must not hold any non-code
files bigger than 100kB this package became useless as the
geometry files are usually bigger than this size. How to
store the files in the future in a reasonable way is not
clear - perhaps in some private repositories(?). In any case
not as part of the ATLAS code base at it is now forbidden.
During the svn to git migration most of the files were
already removed anyway.


Former-commit-id: 751435f051c730fd2248e733d85ebf70614dbdc6