removal of MuonAlignGenTools and all traces
The package MuonAlignGenTools was once implemented to make
track based alignment of the muon spectrometer. The main
developer and at the same time user (Robert Harrington) left
Atlas years ago. Today this package is not at all used for
anything anymore, and nobody knows how to run it or if it
is still runs at all. Therefore the code can now be removed
to not waste time in the migration to MT or whatever. It
will be gone in master and 21.3, but if people are still
interested they could have a look at 21.0 where it will be
kept.


Former-commit-id: 2e60085c003fa6cc17a341a3253d86880a6f0ece