Migrate MuonTrackTruthTool to use handles
MuonTrackTruthTool now uses DataHandles to retrieve from StoreGate.
Previously SG interactions were hidden behind wrapper methods, which seems pointless, so I dropped those methods.
Previously the addSimDataToTree and addCscSimDataToTree methods were passed the key of the SDO container which they used to retrieve it.
Now, the container is retrieved via a ReadHandle in the main body of the code and the method is passed a pointer to it.


Former-commit-id: d9e62148365238c22317b164f05f9453d914d9f6