Updated ART test scripts for MuonRecRTT (ATLASRECTS-4132)
The input files for MuonRecRTT ART test were put in the EOS
directory. But the storage space in EOS is not enough for
the test, so now we access Grid datasets directly. For the
test with pileup, the datasets were produced by the validation
group; For the test without pileup, the datasets were newly
generated.

In addition, according to the requirement of ART framework,
the filenames of test scripts are all turned to lower case.


Former-commit-id: 0e7b3b61dece1aa6b097e7359b228d09d2a62184