Const-correctness fixes for TrigT1CaloTools
Add missing const specifier to some methods to allow them being called
via a const ToolHandle. The only non-trivial change is the removal of
resetting of m_TEMasks in L1EtTools::moduleSums as this was not being
used anyway (the call to fillMaskedOutMap has been removed a while ago
already).


Former-commit-id: edecf3112def82fd852155d851f6cac78c61491d