Migrate MuonTrackPerformance to use DataHandles
This commit migrates several algorithms and tools in MuonTrackPerformance to use DataHandles and also does some code cleanup (removing unused methods and members, etc.).
The main change is that retrieval of objects from StoreGate is moved from the MuonTrackStatisticsTool to the MuonTrackStatisticsAlg.
This necessitated introducing a new method for the latter to call the former by: since other methods were no longer necessary, they were removed.
Changes to MuonPerformanceAlg and MuonTrackPerformanceAlg are more minor, mainly just the migration.
Mentioning ATLASRECTS-4280 for bookkeeping purposes.


Former-commit-id: 5ce1cc55f3e0fe0c813fc3ea83c43f8c069971a4