adjust code to produce SQL files for muon layouts
At the moment the PREsql files, tables and blobs needed to
upload muon layouts to Oracle can only be generated in a
special setup of release 18.9.0 . This modification resolves
issue, however, especially the AGDD portion of the layout
will just be prepared for the DB w/o checking if it really
runs. A subsequent MR will address this flaw.


Former-commit-id: 7885313c8c407b998c1614165205c2d79ddcb358