InDetTrigTrackSlimmer is abusing a DataVector bug to alter constant objects in StoreGate.
A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.

In TrigTrackSlimmer::hltExecute(), we have:

{noformat}
      for (TrackCollection::const_iterator it = m_allTracksFromStoreGate->begin() ;
	   it != m_allTracksFromStoreGate->end()   ; ++it) {

	if (!m_sumTool.empty()) {
	  m_sumTool->updateTrack(*it);
	}
{noformat}

So we call updateTrack() on objects retrieved from StoreGate

This of course needs to be fixed for AthenaMT.
In preparation for trying to finally remove this bug from DataVector,
this has been adjusted in !XXXX so that the const_cast is explicit.


Former-commit-id: 92e0c979f8cc607d54526fecb0378b6825230332