Migrate MuonCaloTagTool to use DataHandles
SG retrievals that previously happened in CaloMuonLikelihoodTool and TrackDepositInCaloTool are now moved to the MuonCaloTagTool.
Moving them to the MuonCombinedInDetExtensionAlg would have been awkward because of the other algorithms that use that code.
In the future, perhaps the MuonCaloTagAlg should have its own algorithm, but this is an acceptable solution for now.
This change necessitates that the relevant containers be passed to the tools, requiring some changing around of internal methods.
Also, a small change is needed in the trigger configuration, since the name of the CaloCellContainer is now set in the CaloMuonTagTool and not in the TrackDepositInCaloTool.


Former-commit-id: 36331cb642b6a6500c8001e6b9f4286c1c449007