LArCalibUtils is abusing a DataVector bug to alter constant objects in StoreGate.
A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.

LArStripsCrossTalkCorrector is retrieving LArAccumulatedCalibDigitContainer's
from SG and modifying them:

```
    currDig->setSampleSum(SampleSumInt);
```

LArCalibShortCorrector also retrievings LArAccumulatedCalibDigitContainer's
from SG and modifies them:

```
 shortedDigits[ii].first->setSampleSum(newSampleSum);
```

These of course won't work with AthenaMT.

(This package also has explicit const_casts in LArDuplicateConstants
and LArRampFCalCorr!)

In preparation for trying to finally remove this bug from DataVector,
adjust so that the const_cast is explicit.


Former-commit-id: 38e136208c8e316dabcfa4fdd22faf4309c70620