TrigCaloRec is abusing a DataVector bug to alter constant objects in StoreGate.
A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.

FullCaloCellContMaker copies const cell pointers from an existing
collection into a new one.  This in turn is called from TrigCaloCellMaker,
which can then run correction tools on the produced collection,
modifying the cells in SG.

These of course won't work with AthenaMT.

In preparation for trying to finally remove this bug from DataVector,
this has been adjusted so that the const_cast is explicit.


Former-commit-id: e4c8ae4446919aec8bedb0941fa7e530fe5a12eb