MuGirl is abusing a DataVector bug to alter constant objects in StoreGate.
A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.

MuGirlRecoTool::MuGirlReco a takes const InDetCandidateCollection
and modifies it via addTag.

This of course won't work with AthenaMT.

In preparation for trying to finally remove this bug from DataVector,
this has been adjusted so that the const_cast is explicit.


Former-commit-id: c5c1f5e096593d48c600e3099a876e9674a9a13f