adding back NSW sensitive detectors to digitization output
Probably by accident the NSW sensitive detectors (aka simhits)
are not longer propagated to the digitization output. This is
however a basic requirement to run fast digitization what is
currently still needed. This MR adds them back. They should
not be removed w/o checking with the NSW developers list.


Former-commit-id: a2422cb1868a9572aaf831542024870d96b44ac0