TileRecAlgs is abusing a DataVector bug to alter constant objects in StoreGate.
A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.

TileRawChannelToTTL1 retrieves a TileRawChannelContainer from SG
and then possibly modifies it.

This of course won't work with AthenaMT.

In preparation for trying to finally remove this bug from DataVector,
this has been adjusted so that the const_cast is explicit.


Former-commit-id: dca90e4b111a05d8fa351b73b5be43b190900be6