TileRecUtils is abusing a DataVector bug to alter constant objects in StoreGate.
A long-standing bug in DataVector has meant that it was possible to get
a non-const pointer back from a DataVector::const_iterator.

TileCellBuilder retrieves a TileRawChannelContainer from SG and then
possibly modifies it.
There are also issues in TileRawChannelBuilder, TileRawChannelNoiseFilter,
TileRawChannelOF1Corrector

This of course won't work with AthenaMT.

In preparation for trying to finally remove this bug from DataVector,
this has been adjusted so that the const_cast is explicit.


Former-commit-id: ad34e9c47ffc249956f2716c403e32a9fa660f9c