Changes to resolve conflicts in the master branch and bring up to date with essential developments in the 21.X branches
These are changes to resolve many of the conflicts that were unfortunately
erroneously introduced by *purely cosmetic* changes in the master branch,
when that actual main development branch was 21.1-dev - so the changes
should have gone into that branch.

This means that many essential, and well tested functional developments
and bug fixes to files in the 21.* branches can no longer be swept into
master.

These tags partially fix the conflicts, while also porting *all* the
essential bug fixes and developments, while also implementing some of
the minor cosmetic changes.

Not all the cosmetic changes have been icluded - notably those which
change the names of class variables etc have not been included in
order to limit any potential sources of difference in the validation
tests. Once the actual essential functional changes have been validated
and there is a stable set of validation results for comparison, then
the purely cosmetic changes to the class structure can be attempted,
but not before.


Former-commit-id: 9306f1ccad81ebb7d57446f3a5e51c2b0a1a1d15