Migrate TrackSegmentAssociationTool to use DataHandles
The TrackSegmentAssociationTool now retrieves segments using a ReadHandle.
Additionally, the tool was not attempting to associate segments to muons reconstructed by the MuGirlLowBeta algorithm.
I moved this decision to the MuonCreatorTool and added a property to control it, to make it easier to change this later if so desired.
Also, unused ToolHandles are now disabled.  Mentioning ATLASRECTS-4300 for bookkeeping purposes.


Former-commit-id: a463596669d4a203479ed62e49a0b2e8ba4aebbc