MCTester_i: DataVector const workaround
DataVector has had a long-standing bug in which it was possible to get a
non-const pointer from a const_iterator.  This package was relying on this
to compile.  To fix this, we need to use a const_cast.  Can't easily
get rid of it, since it stems from a problem in the external package.



Former-commit-id: 5bb8144d67b901e20c1eaaf433cec3964cefa527