AthContainersInterfaces: Use ConcurrentBitset for auxid_set_t.
We have been using a std::unordered_set to represent a set of auxid_t.
This is not a thread-safe data structure, however, so we've often had
to maintain thread-local copies of these sets.
But the total number of auxid_t's is not too large --- less than 2000
for a typical reco job.
In fact, performance tests show that for auxid_t distributions typical
of a reco job, a ConcurrentBitset has better performance than
std::unordered_set for pretty much all operations that matter.

So change auxid_set_t to a ConcurrentBitset.



Former-commit-id: 459ee6febb49a2995ef00fedea5e4f121d81afbf